renamed makeboard to pandemicgame (makes more sense)
